Support daily dry run release (#3548)
## Motivation and Context
Runs dry-run release workflow daily

## Description
We've had quite a few instances where we discovered that a dry-run
failed when we were about to release, by which time a good amount of
code changes were accumulated in the `main` branch. That sometimes made
it harder to investigate the underlying root cause, compared to if we
had run dry-run daily. To alleviate the issue, this PR runs a dry-run
release daily, ensuring that the `main` branch is in good shape to kick
off a prod release and that we can react to a dry-run failure caused by
a PR merged to the `main` the previous day.

To make this happen, the existing release workflow `release.yml` has
been converted to a reusable workflow (with `workflow_call`). This is
because a scheduled workflow run cannot take inputs, making it difficult
to pass `commit_sha` and the `dry_run` flag to it. With `release.yml`
being a reusable workflow, we have two new workflows calling
`release.yml`: `prod-release.yml` and `dry-run-release.yml`, both of
which we can manually trigger and we can also trigger the latter via
cron.

Note that there is no longer a checkbox `dry-run` that used to exist in
`release.yml`. Instead, we choose a corresponding workflow.

## Testing
Verified the previous workflows continued to work:
- manually triggered dry-run release
- manually triggered prod release

However, a scheduled dry-run has not been tested because we first need
to check-in `dry-run-release.yml` to main for a scheduled workflow to
kick-in. In other words, we'll do live test and see what happens (will
fix if any issues come up).

----

_By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice._