Fix several waiter path matcher issues (#3593)
While implementing waiters in another branch, I discovered and fixed a
number of issues with the path matcher codegen logic. These issues were:

- Generated code for nested flatten projections failed to compile due to
the first projection producing a `Vec<Vec<&T>>` instead of `Vec<&T>`.
- Path matchers that don't use input were taking input as an argument
anyway, which results in an unnecessary clone of the input when used by
the generated waiter logic.
- The comparisons generated by `RustWaiterMatcherGenerator` would fail
to compile whenever comparing a string against an enum.

This PR fixes all these issues.

----

_By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice._