Loosen s3 integration tests (#3641)
## Motivation and Context

Updates integration test(s) for a pending S3 model update that would
affect one of the query parameters. Where possible I've kept as much of
the assertions/spirit of the test as possible as opposed to fully
ignoring/disabling it until these updates are available.

Same as https://github.com/smithy-lang/smithy-rs/pull/3628

Downloaded pending model changes and ran tests against it and current
version of model to see that they pass in both states.