Renaming gradle task `:aws:sdk:test` to `sdkTest` avoid task name collision (#3650)
It is now called :aws:sdk:sdkTest and the README has been updated to
align with this.

## Motivation and Context
<!--- Why is this change required? What problem does it solve? -->
<!--- If it fixes an open issue, please link to the issue here -->
The gradle task `:aws:sdk:test` mentioned in our `README` has been
broken for sometime. This seems to be because one of the plugins we are
using (I was unable to determine exactly which one) is registering a
task named `test` and it was conflicting with ours.

## Description
<!--- Describe your changes in detail -->
I changed the name of the broken task to `:aws:sdk:sdkTest`, explicitly
registered it, and added the appropriate dependencies and steps for the
task. I also updated the `README` to reflect this new name and added a
note to indicate that these tests require Go to be installed (because
they tests some FIPS functionality).

## Testing
<!--- Please describe in detail how you tested your changes -->
<!--- Include details of your testing environment, and the tests you ran
to -->
<!--- see how your change affects other areas of the code, etc. -->
I was able to successfully run `./gradlew :aws:sdk:sdkTest` on my local
machine (after installing Go which I was initially missing).

## Checklist
<!--- If a checkbox below is not applicable, then please DELETE it
rather than leaving it unchecked -->
No changes to any published code in this PR
----

_By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice._